+++ coursecake/flaskapp/limiter.py
@@ -1,0 +1,7 @@
+from flask_limiter import Limiter
+from flask_limiter.util import get_remote_address
+
+# enforces rate limits for ALL routes
+limiter = Limiter(
+    key_func = get_remote_address,
+    default_limits = ["1/second; 20/minute"])
